fixed polygons according to spec
According to spec and uses of most GIS applications/mapping frameworks like OpenLayers3, a polygon is a array of arrays of points. 

I am unsure if its a error in the documentation or you omited this also in the implementation parts (I know its just storage,  but if querying tools and helpers are unaware I assume they break ).

You may use the Pull Request as you see fit.