Take care of the case when server side returns status 400 and requests to decrease page size. (#421)
Take care of the case when server side returns status 400 and requests to decrease page size. This can happen in the following case (not a typical case):
(1) There was at least one LSN that fits into response.
(2) One of further LSNs is of size > 1 (more than one document), fist doc is fitting and some other doc is not fitting.